Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Ah, I see. I would expect that the fan should turn on upon transition to S0 (if needed) and then turned off upon entering S5. Therefore, in the case that you posted, the fan shouldn't already be on.

Yes, maybe I should be fixing it so that the fan doesn't turn on. The default state for the fan when the system comes up is on, and as it happens, nothing is turning it off.
Maybe on most systems the initial power-up sequence turns the fan off initially, and I will have to work through the initial start-up sequence to see where the fan should be turned off.

Patch-set: 1