Update patch set 4
Patch Set 4:

Let's make kukui define a default bootblock size (of 20992).

If a boot block is provided, assert that it is less than or equal to default size (ensures that an update to the bootblock size in coreboot is detected)

If boot block is not provided, then use dummy bootblock of size default.

Patch-set: 4
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>