cr50: make DIOB7 an interruptable and wakable pin for EC-EFS2
Per the design in go/ec-efs2, EC shall drive DIOB7 high or low
to inform CR50 a packet mode start or end. When EC drives it high,
CR50 should wake up if it was in a deep sleep.

This CL adds interrupt handlers on DIOB7, and sets it up as a
wakepin during CR50 initialization time.

BUG=b:141143112
BRANCH=cr50
TEST=None

Change-Id: Id68d13df0e3b6db25e355dbea653e6e1fce145c6
Signed-off-by: Namyoon Woo <namyoon@chromium.org>