Servo Micro: Temporary solution to SPI flash failures on Servo Micro
Temporary removes the disable of the SPI peripheral which is
performed after each SPI transaction on the Servo Micro platform.
This improves the reliability of the SPI transactions with the flash
chips leading to few failures on Flashrom.

Under normal situations, about 1 in 20 or so reads from a 16MB flash
chip will return an incorrect payload although the frequency is
inconsistent.

BRANCH=servo
BUG=b:144846350
TEST=Manual Testing with the following process:
Setup loop to verify the flash image or read the flash image
and perform a binary comparison against a the known image.
It takes about 100 cycles to ensure that the problem is resolved.

Change-Id: I5fec4dd74ece6394d68157999ec8a55947185046