Update patch set 2
Patch Set 2:

> I wonder if it would be better to tell charge manager about barrel jack current once the AP is off. I believe once charge manager is informed of the power, it will then make the decision to switch suppliers.

That's right, but the switch is prevented unless the AP is off. We'd want a mechanism to defer the charge update until next AP shutdown.

I think skipping trying that is okay- it would be rather involved to support, and there isn't any user feedback on what power source you're actually using anyway. It's possible a user might connect a new power source to get better performance that we can't switch to immediately, but there's no feedback so the only way to ensure it actually switches power source is to disconnect the old one and in that case charge port selection will run again.

Patch-set: 2