Update patch set 1
Patch Set 1:

The I2C 7-bit address for this device is 0x08 and not 0x0A.  Given that, it works some of the time.  The mechanism used to wake this device up is not very deterministic.  It wakes some of the time and not at others.  You should also add some code to verify the power rail is up for this device before trying to access it so you remove the "2019-12-23 10:23:44 [172.474571 Err: set retimer port(1): 1]" lines.

Patch-set: 1
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>