USB PD: Consider DFET of battery when refrain from snk_reset
CL:1565147 introduced a new clause to the SNK_DISCOVERY state to
prevent reset timers from starting when a battery is reporting below a
configured level, which indicates the point at which the battery is
capable of withstanding the potential loss of Vbus.

There is an another case of having good battery level but battery can't
discharge due to the DFET status of fuel gauge so we need to consider it
as well.

BUG=b:146758440
BRANCH=octopus
TEST=Verify on the DUT with bad batterh which can report battery level
correctly but DFET is disabled and can't be revived.

Change-Id: Ia3f3d26625a054f631fdddc74065f2e829852989