cr50: make DIOB3 interruptible and wakable for EC-EFS2
This CL enables both rising/falling-edge triggerred interrupt on
H1 DIOB3 pin and makes it wakable as well, only if EC-EFS2 is
supported.

Cr50 reads the DIOB3 level with internal pull-up on it, and detects
if the board supports EC-EFS2. Such boards have an 1M pulldown
registor on DIOB3 and others have an 10.

If EC-EFS2 is supported, Cr50 enables interrupts on DIOB3 and
reconnect GPIO_AP_FLASH_SELECT from DIOB3 to DIOB4.

Before CR50 deep sleeps, it will set the DIOB3 as wakable on rising
edge, and stores the EC board_mode into a power-down scratch register.
On wake-up, it recovers EC board_mode value from that register.

BUG=chromium:1035706
BRANCH=cr50
TEST=None

Change-Id: I1221a1a19219274622ab710568ce7c66ab2f1da7
Signed-off-by: Namyoon Woo <namyoon@chromium.org>