Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I don't understand what tcpci_nct38xx_get_cc() was trying to do.
> > There is this comment from when it was added:
> > https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1436658/4/driver/tcpm/nct38xx.c#169
> > "tcpci_tcpm_get_cc() does not behave correct when we set the pull resistor (Rp or Rd) only to one CC line (as I explain in my comment on set_cc function), therefore we had to override it with this function."
> > But I can't understand how tcpci_nct38xx_get_cc() solves the problem.
> > Amit / CH, can you explain more, or show me which part of the spec this is following?
> 
> The problem is that the original code in tcpci_tcpm_set_cc() set the Pull resistor (Rp or Rd) to both CC lines. 
> This is not compatible with TCPC spec which say that the pull value (Rd or Rp) should be applied only to one CC line (according to the connection polarity) and the other CC line should stay Open. Setting both CC lines to the same value will result in a wrong CC status and wrong connection handling.
> So once we override  tcpci_tcpm_set_cc() with tcpci_nct38xx_set_cc() , we started to get wrong result from tcpci_tcpm_get_cc() since this function assumes that both CC lines have the same Pull value. Therefore, we also needed to override  this function.

From my understanding of the old nct38xx and tcpci code, get_cc works exactly as it did in the before, as long as DRP is being used.  Which is when we have power to the AP.  When the power is not enabled on the AP, we do not look like a DRP connection and only look for a charger to get power.  This is where the nct38xx driver used the ROLE to give the extra information needed to get the correct CC values.  This is due to CC_STATUS CC State fields both being 0 if Looking4Connection is 1.

Patch-set: 4
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>