Update patch set 5
Patch Set 5:

My understanding is that this is due to the behavior of the CC_STATUS.ConnectResult  
The original code was using ConnectResult to determine the CC1/CC2 value.
When ConnectResult  = 0 it mean that the TCPC is presenting Rp 
When ConnectResult  = 1 it mean that the TCPC is presenting Rd 

However as long as both CC1 and CC2 in ROLE_CONTROL register are set to Rd then ConnectResult  represent the correct value. But once a connection is detected the TCPM firmware changes CC1/CC2 in ROLE_CONTROL according to the cable orientation (one CC is Open and one is Rd) and in this case when only one CC line is Rd the ConnectResult is set back to 0. Therefore the ConnectResult by itself cannot be used as an indication and we also need to read ROLE_CONTROL value to check if the TCPC is presenting Rp or not.

The TCPC spec in the CC_STATUS Register Definition specifically writes that ROLE_CONTROL should be checked. Here is the relevant text:
=============================
If (ROLE_CONTROL.CC2=Rp) or (ConnectResult=0) 
00b: SRC.Open (Open, Rp) 
01b: SRC.Ra (below maximum vRa) 
10b: SRC.Rd (within the vRd range) 
11b: reserved 

If (ROLE_CONTROL.CC2=Rd) or (ConnectResult=1) 
00b: SNK.Open (Below maximum vRa) 
01b: SNK.Default (Above minimum vRd-Connect) 
10b: SNK.Power1.5 (Above minimum vRd-Connect) Detects Rp 1.5A 
11b: SNK.Power3.0 (Above minimum vRd-Connect) Detects Rp 3.0A 
=============================

Patch-set: 5