Waddledee: Initial board file
This commit contains the initial code for waddledee, the ITE reference
board for dedede.  All of the GPIOs are defined, and any tasks that can
be enabled to begin with are.  Since the charger driver is still in
progress, this commit undefs any charger related baseboard features,
and they will be re-enabled once the driver is ready.

BUG=b:146557556
BRANCH=None
TEST=builds

Change-Id: Ie5901304f92bf9040f1c883fb738c9f3bda0e95e
Signed-off-by: Diana Z <dzigterman@chromium.org>