bloonchipper: Add transport and sensor detection
bloonchipper (aka hatch_fp aka dragonclaw) has a voltage divider that
can be used to select the sensor and the transport type.

Supported designs:
* Dragonclaw rev 0.2 (green with Google logo):
  go/dragonclaw-schematic-rev-0.2
* Hatch reference v3.0:
  go/hatch-schematic-rev-3.0

The selection lines are connected to ADC inputs, so a future change will
use the ADC to allow more than two transports or sensors.

BRANCH=none
BUG=b:147113851
TEST=flash dragonclaw rev 0.2 and view console output

Change-Id: If2e4b150d34cfe41477be528c70e1645043d4d82
Signed-off-by: Tom Hughes <tomhughes@chromium.org>