stm32h7: Fix flash protect reset loop
This patch fixes the reset loop that would occur
when software write protect (ro_now) was enabled,
but hardware write protect was disabled.
The reset loop would be triggered by a reboot of
MCU.

The issue is that flash_pre_init wants to remove ro_now
if hardware wp is deasserted on boot. It then assumes
that hard resetting the MCU could clear this flag.
Unfortunately, this is not the case, so the MCU ends up
in a never ending reset loop.

This patch still allows the MCU to try a hard reset once,
but will accept defeat if ro_now is still enabled on the
second boot.
The only reason we still leave this reset here is
because there might be a scenario where a soft reset
occurred moments before calling commit_optb, but after
setting write protect for some flash regions.

BRANCH=nocturne,hatch
BUG=b:146428434
TEST=Check on Nucleo-h743zi

Change-Id: Ie93d9a968ed03caf00430df02a514c0cf1da9f26
Signed-off-by: Craig Hesling <hesling@chromium.org>