Update patch set 12
Patch Set 12: Code-Review+2

Ok, I see now, thanks for the explanation, and making the code clearer.
It seems like a lot of work just to decide if we are Source or Sink, but I guess it is what we have to do.
It is also slightly scary to imagine cc1_present_rd != cc2_present_rd in the non-DRP case.
On the plus side, it looks like we have reached almost the same implementation as mt6370_get_cc().

Patch-set: 12
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2