nocturne_fp: Add fix for legacy reset loop
This patch mitigates an infinite reset loop caused by an RO bug.
The reset occurs in RO when hardware write protect (wp_gpio_asserted)
is disabled, but software write protect (ro_now) is still enabled.
This can be seen by disabling hardware write protect and issuing
a soft reset.

There is one case where RO will forgo issuing this system reset.
That is when it detects a power on reset. Furthermore, it retrieves
its reset flags from the main system_get_reset_flags function, which
combines hardware reset registers AND a special RTC backup register
designed to preserve reset flags.

We exploit this reset backup register mechanism to inject a fake power-on
flag before resetting. As an added bonus, we also inject an ap-off flag
so that we can determine on startup if the power-on flag is real
or forged by this mechanism. If we detect that the power-on flag was
forged, we print a warning and fix the current reset flags.

In order to ensure that a power-on will be forged when
a spurious reset happens (exception or pin reset), we keep
the backup register loaded with the power-on and ap-off reset flags,
when the hardware write protect is disabled.

In order to keep the typical code path (HW+SW WP enabled)
clear of complexity and false power-on reports, we only forge
the power-on flag when hardware write protect is disabled.
Thus, we conditionally setup the forge on startup and setup an
interrupt handler to catch changes to the hardware write protect status.

It is safe to use ap-off flag for our nefarious purposes, since
the fingerprint controller has no functionality to control an AP
and has no included code that uses this reset flag.

Review:
Normal power on reset --> The ap-off flag should be cleared
Forged power on reset --> We set the ap-off flag

Scenarios covered:
* True power on      --> No reset loop and ap-off would not be set
* HW reset pulse     --> We preloaded ap-off and power-on flags in
                         the reset backup register
* Exception/Watchdog --> Same as above
* System reboot      --> We modified the system_reset function to
                         add ap-off and power-on to reset backup
                         register

BRANCH=nocturne,hatch
BUG=b:146428434
TEST=make buildall -j
TEST=Checked all of the scenarios mentioned above
     in the [SW-WP off + HW-WP off], [SW-WP on + HW-WP on],
     and [SW-WP on + HW-WP off] situations using the nucleo-h743zi
     board (FILL IN WITH CL OF NUCLEO PORT).

Change-Id: I89361fa95be8eafe78c80c30f5b3195d7a724f81
Signed-off-by: Craig Hesling <hesling@chromium.org>