Makefile: Add print-make-vars rule
The Bash-based flash_ec script uses print-baseboard to identify if a
board has a baseboard.

The newer Python version I'm working on not only uses that
information, but reduces the number of board-specific quirks we have
encoded in the script by using the CHIP, CHIP_FAMILY, CHIP_VARIANT,
and CORE variables to discover quirks. Introduce a print-make-vars
rule to be used by the new flash_ec script which is extensible to many
variables.

BUG=b:145292630
BRANCH=none
TEST=make BOARD=... print-make-vars for a few different boards

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ib91123d2b0b36c584db563e192f72a3ac001f330