Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> If we want to support MODE0, then I think we should add pull down resistors to the stm side of the level shifter.

That would require a very strong pull-down, which would also impact the ability to drive high on clock.  The STM32 already has to contest with a pretty strong pull-down on the far side of the mux and is barely able to meet Vih for the full clock half-period, so anything that makes it harder to drive high may make the clock unreliable entirely.

> 
> What ever FW hack we get to work on this STM32 might not work with timing on other stm32 versions. The external pull down adds much less overall complexity to the c2d2 design.

The feature can be #ifdef'd for the F0 if that's the main concern.
However, as noted in the next comment, I think we can avoid creating a race condition by simply keeping SPE enabled.  I can try playing with that approach (or you can if you have some time today).

Patch-set: 1
Reviewer: Gerrit User 1002022 <1002022@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>