Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What are the downsides that you have seen with this? (hopefully none)
> 
> I think I want to so something similar for the new implementation of c2d2 (stm-based debug board replacing servo micro).
> 
> We are having issues with the clock signal not being driven by the stm32 when the SPE bit is cleared. When the clk signal is undriven there are pull ups that cause a rising edge on the CLK signal which cause erroneous data.
> 
> We want to leave the SPE bit set more than it is today for C2D2. I haven't tried this solution yet, but this might be what we are looking for in general.

I have seen no regressions in functionality and based on the email regarding it's status when SPE is disabled I'd be fully in support of eliminating all toggling of the STM32_SPI_CR1_SPE bit during the flash operations and revert it to the prior behavior before this CL.

https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/853715/

Before we were turning on the STM32_SPI_CR1_SPE when initialization occured and only turning it off on shutdown. Right now we're turning it on and off twice during each transaction which doesn't appear to follow the procedure in STM's reference SDK or the reference manual.

Patch-set: 2
Reviewer: Gerrit User 1347380 <1347380@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0