Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> CCing Brian since https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1979418 makes it look like servo micro is running into the same problem

I've done some tests with the calls to disable STM32_SPI_CR1_SPE outside of the spi shutdown eliminated on Servo Micro and have not had issues after over 30 cycles. Does leaving SPE on apart from the shutdown eliminate the glitch? Older versions of this file did not manipulate the enable and disable control as aggressively until we added the half duplex operating mode.

I have not done tests which exercise half duplex and we haven't done any power measurements to see the delta so we may not want to fully revert it yet.

Patch-set: 1