Update patch set 3
Patch Set 3:

"we have a hardware line to this pin": I'm not seeing it on my schematic, is that with a re-work? Is there a bug for that HW change? Why not set this via I2C instead of changing the HW?

Patch-set: 3
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>