Update patch set 1
Patch Set 1:

Some progress with the nucleo-f767 board...

I have read thru some more of the code to understand some of the board startup/various modules.  
And was able to port Vincent's 2017 nucleo-f767 to the tip of master (with some hacky changes).  

Note: I did clean up all of the pre-submit checks style checks, with the exception of the "make buildall -j" step.  
Some of the boards appear to have compilation failures, but it also seemed to fail when I tried it on master.  
So at the end I did 
  repo upload .  --cbr --re=nsanders@google.com --no-verify
to at least share the code with others.  


I'm thinking the next step is probably to start experimenting with USB (I'm guessing the simplest thing is to try to get it to show up as a device when attached to a PC, with just the usb_console running?).  

And then to start trying to get the USB<->UARTS working...

#### USB ####
From some poking around, I think the _dwc_ versions of the USB code are the right ones (i.e. synopsis designware).  From some skimming of the STM32F4 and STM32F7 reference manuals the USB blocks look "pretty similar".  

Sadly this doc:
  AN4660 Application note Migration of microcontroller applications from
  STM32F42xxx/F43xxx devices to STM32F7 Series devices
Says: "Configuration – SW not compatible" for "2.15 USB OTG".  
The only obvious changes I see are additional endpoints in the STM32F7, so hopefully it isn't too different.  

Also, there is only a build in phy for usb FS speed (12 Mbit).  
To get to FS (480 Mbit) we would need an external ULPI phy chip, and also to program the USB_HS core.  
I did see some references to ULPI in the code, so I am hoping most of the code is there...

#### FUTURE ####
There is existing code for things like SPI, I2C, ADC's, etc.  
* But, AFAIK there isn't any code for USB-JTAG (as a fallback we can put on an external connector so a Segger/ST-Link can be used).  
* Also, there isn't any ethernet related code (for future capability of ethernet -> uart).  But that is an optional/future feature at this point.

Patch-set: 1