usb pd: Don't force VSafe5V when there is no active charge port
This CL fixes an issue where USB PD would not allow the max request
when there is currently no charge port. This was causing issues
following the jump to RW when the PD contract gets
reestablished. There is a race condition between the PD contract
negotiation and the charage manager being updated from the first 5V/3A
state. This race condition results in selection of the vSafe5V PDO and
this can cause problems with external adapters resulting in a hard
reset. Note that this problem only occurs when no battery is present
since otherwise there wouldn't have had a pd contract negotiated when
the EC is executing from RO.

BUG=b:145783611
BRANCH=firmware-hatch-12672.B
TEST=On Kohaku using the kohaku provided external adapter, verified
that kohaku can boot to the OS when SW sync is enabled and no battery
is present.

In addition, connected two kohaku chargers simultaneously (via power
strip on button) and verified that only one port negotiates VBUS at 20V.

Change-Id: I68b21f2a86634d4bae91a5b812f2a61b302b0ef3
Signed-off-by: Scott Collyer <scollyer@google.com>