common: make ARRAY_SIZE, ARRAY_BEGIN, ARRAY_END type safe
ARRAY_SIZE (and family) used to allow pointers to be passed to the
macro, which would cause a value of one for the array size. This is a
really easy accidental programming error.

This change causes ARRAY_SIZE and ARRAY_END to divide by zero (using
BUILD_CHECK_INLINE) if a pointer is passed. ARRAY_BEGIN will cause a
BUILD_ASSSERT to fail if a pointer is passed.

BUG=none
BRANCH=none
TEST=buildall

Change-Id: Ibdd0efcc9fc46c43d46f2bd06529cc76df910f2e
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>