Update patch set 2
Patch Set 2:

I'm at the point where USB is "starting to work".  
The code compiles, and when I run "usb on" and connect the OTG_FS port to a linux machine it starts to enumerate.  
It fails to complete enumeration, though.  
If you have suggestions on where to look that is welcome.  

My first guess is that the clock speeds/timing between the F767 and F446 (in sweetberry) might be
different enough to need some adjustments.

Patch-set: 2