Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> +2 to stop the crash, but why are we overriding svdm_dp_post_config and not using usb_common.c?
> Hopefully we can move to using the common version.

This is the code as it was.  I'm just trying to get rid of the crash.  I pulled this zork specific svdm code out and it seemed to continue to work properly with USB-C0 and continued to not work with USB-C1.  When we get USB-C1 working, let's consider using common or using common with some CONFIG flags to do the extras that we also need.

Since I am currently tweaking the zork specific code, I am going to "fix" the usage of hpd_deadline as well in this CL

Patch-set: 2
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0