kohaku: Define CONFIG_MKBP_EVENT_WAKEUP_MASK as 0
This allows the AP to set MKBP wake masks, which will enable the EC to
wake the system from suspend states.  The default is 0, which means that
the AP gets to choose which events will wake it from suspend.

BUG=b:144122000
BRANCH=firmware-hatch-12672.B
TEST=Verify 'ectool mkbpwakemask set event 0x400' succeeds.
Also verify with corresponding coreboot CL that the system can wake
from suspend when a DP-capable monitor is plugged in.

Change-Id: Ife726f341770e6c9cea2ffe39abc9d8ed38b0cbb
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>