Trogdor: Abandon virtual HPD approach and configure TCPC for HPD status
The TCPC is required to know the HPD status; otherwise, some mux
misbehaves. Even thought the HPD status is through a GPIO to notify AP,
still configure the TCPC.

As the GPIO approach works well, abandon the virtual HPD approach.

BRANCH=None
BUG=b:147358149
TEST=Plugged a Type-C to DP dongle to the board, checked that the IN_HPD
bit not set. Plugged a monitor to the dongle, checked that the IN_HPD
bit set.

Change-Id: Id8b8eae462c80c36156a1e9443c8af87cd784fd0
Signed-off-by: Wai-Hong Tam <waihong@google.com>