tcpm/raa489000: Revise the init function
The datatsheet for the RAA489000 talks about an unlock sequence that
should be done after POR.  This commit adds the unlock sequence to the
init.  Additionally, a number of ADCs and comparators are enabled such
that we can actually detect CC_STATUS changes along with VBUS presence.

BUG=b:147316570
BRANCH=None
TEST=build and flash on waddledoo, verify that we can read CC status and
VBUS.

Change-Id: I685febaa13c54462055909d31f95b1f8f9199eb0
Signed-off-by: Aseda Aboagye <aaboagye@google.com>