waddledoo: Add Type-C, PD, and charging support
This commit adds USB Type-C, Power Delivery and charging support to
waddledoo.  Waddledoo is using the new TCPMv2, low power mode has been
disabled for the time being while we sort out issues with the TCPC.
Currently in this state, we cannot actually receive any PD messages, but
the PD task is running.

BUG=b:147257992
BRANCH=None
TEST=Build and flash on waddledoo, verify that we can detect Type-C
chargers, and that it can charge a battery.

Change-Id: Id30218ef685ed27a934bae3a580f47754f659ac6
Signed-off-by: Aseda Aboagye <aaboagye@google.com>