Charger: Create charger driver structure
With upcoming boards which use multiple charger chips, the EC codebase
needs to be changed to assume chargers may have different I2C ports.
This commit creates the driver structure and wrapper functions, which
for now are hard-coded to chip 0 for equivalent behavior with previous
code.  A general charger config is created for all boards in charger.c
for now, which uses the build information to fill in the structure.

For drivers with non-static and non-driver exports (bd9995x, rt946x),
the approach was to code static functions to expect a charger chip
number, while having non-statics hard code that value to 0.

For boards which may be supporting different I2C configurations for the
charger over board versions, they may define
CONFIG_CHARGER_RUNTIME_CONFIG to fill in these fields after boot.

BRANCH=none
BUG=b:147672225
TEST=builds, chargers on hatch and octopus work

Change-Id: I390ede494226252e512595c48099fa1288ffe93e
Signed-off-by: Diana Z <dzigterman@chromium.org>