Update patch set 2
Patch Set 2:

> > wdyt? or just the for-loop?
> 
> I like the suggested way of doing it because it ensures the lowest possible overhead.
> 

you mean for-loop, right?
this i_++ in the for-loop, what asm code is it converted to, does it deal with memory store/load or with registers? can we do just some register ops with more predictable timing (or NOP? but we still want the counter for the loop, so there will be register ops anyways)? or am I just being too paranoid? :)

Patch-set: 2
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>