Update patch set 1
Patch Set 1:

Before we can accept this change we will want to test the code with the library. We are working on setting up the private repo for it. In the meantime, can you send it to us for testing?

Patch-set: 1
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>