zork: PS8802/PS8818 dynamic detection cleanup
Changed the method for detection to a non-destructive
mechanism so it could be called from any of the
interfaces.

Added in switching to direct calling the detected
driver once the hardware has been determined.

The 8802 is the main mux when it is present and the fp5 mux
should only select the lanes to send.  Flip should not be
sent to the FP5.  It is a specialized form of a secondary
MUX and currently being placed in as a retimer.  In the
future the retimers will all become MUXes and they will
chain... just not today.

BUG=b:147428570
BRANCH=none
TEST=verify USB-C1 device with AP running

Change-Id: I6b0eedd1dfcc91c3114f8dc481f5ca2841eb9e85
Signed-off-by: Denis Brockus <dbrockus@chromium.org>