cr50: add support for using a strap pin as a ccd gpio
New boards can use the 5kPU strap as a ccd gpio. This change makes DIOA9
a gpio you can assert on boards with the 0x0E strap.

This change uses 2 bits from the board properties to determine which pin
is used as the ccd gpio.
    0 - no ccd gpio
    1 - DIOA1
    2 - DIOA9
    3 - DIOA12
DIOA6 is another strap pin, but there's only one valid strap with a 5kPU
left, so I decided not to use another board property bit to support it
as a possible ccd gpio. I want to save the board property bit, since
we're running out of them and there are so many other I2C straps boards
can use. We can add it later if we need to.

BUG=b:147812066
BRANCH=cr50
TEST=manual. Use pinmux and gpiocfg to verify the output is only enabled
when the gpio is asserted.
     no added brdproperties - nothing is different with pinmux

     run on Puff
     	gpioset CCD_REC_LID_SWITCH 0
	EC shows recovery button pressed
     	gpioset CCD_REC_LID_SWITCH 0
	EC shows recovery button released

     add BOARD_CCD_REC_LID_PIN_DIOA1 to SPI board
     	pinmux output adds
	DIOA1 27  IN GPIO1_GPIO10
	GPIO1_GPIO10 24  DIOA1

        gpioset CCD_REC_LID_SWITCH 0
	gpiocfg shows "GPIO1_GPIO10: read 0 drive 0"
        gpioset CCD_REC_LID_SWITCH 1
	gpiocfg doesn't show GPIO1_GPIO10 as an output

     add BOARD_CCD_REC_LID_PIN_DIOA9 to SPI board
     	pinmux output adds
	DIOA9 27  IN GPIO1_GPIO10
	GPIO1_GPIO10 16  DIOA9

        gpioset CCD_REC_LID_SWITCH 0
	gpiocfg shows "GPIO1_GPIO10: read 0 drive 0"
        gpioset CCD_REC_LID_SWITCH 1
	gpiocfg doesn't show GPIO1_GPIO10 as an output

     add BOARD_CCD_REC_LID_PIN_DIOA12 to I2C board
     	pinmux output adds
	DIOA12 27  IN GPIO1_GPIO10
	GPIO1_GPIO10 13  DIOA12

        gpioset CCD_REC_LID_SWITCH 0
	gpiocfg shows "GPIO1_GPIO10: read 0 drive 0"
        gpioset CCD_REC_LID_SWITCH 1
	gpiocfg doesn't show GPIO1_GPIO10 as an output

Change-Id: If74385135a572e7e5d0763fad9f5368fdec8d7a0
Signed-off-by: Mary Ruthven <mruthven@chromium.org>