cr50: add support for using a strap pin as a ccd gpio
Boards are going to use the 5kPU strap as a ccd gpio. Depending on the
board it will be connected to the lid open or the recovery signal. There
are 4 strap pins. This change restricts the pins that can be used for
this ccd signal to three of them, so we can encode the information in 2
bits. The pins that can be used are DIOA1, DIOA9, and DIOA12. DIOA6 is
the pin that can't. There is only one valid strap with a 5kPU on DIOA6,
so this decision doesn't remove many strap options. If we want to add
that later, we can. It will just take 1 more bit, and we don't have that
many bits left in board properties.

This change doesn't enable the GPIO on any boards. That will be done in
a followup cl.

BUG=b:147812066
BRANCH=cr50
TEST=manual. Use pinmux and gpiocfg to verify the output is only enabled
when the gpio is asserted.
     no added brdproperties - nothing is different with pinmux

     add BOARD_CCD_REC_LID_PIN_DIOA1 to SPI board
     	pinmux output adds
	DIOA1 27  IN GPIO1_GPIO10
	GPIO1_GPIO10 24  DIOA1

        gpioset CCD_REC_LID_SWITCH 0
	gpiocfg shows "GPIO1_GPIO10: read 0 drive 0"
        gpioset CCD_REC_LID_SWITCH 1
	gpiocfg doesn't show GPIO1_GPIO10 as an output

     add BOARD_CCD_REC_LID_PIN_DIOA9 to SPI board
     	pinmux output adds
	DIOA9 27  IN GPIO1_GPIO10
	GPIO1_GPIO10 16  DIOA9

        gpioset CCD_REC_LID_SWITCH 0
	gpiocfg shows "GPIO1_GPIO10: read 0 drive 0"
        gpioset CCD_REC_LID_SWITCH 1
	gpiocfg doesn't show GPIO1_GPIO10 as an output

     add BOARD_CCD_REC_LID_PIN_DIOA12 to I2C board
     	pinmux output adds
	DIOA12 27  IN GPIO1_GPIO10
	GPIO1_GPIO10 13  DIOA12

        gpioset CCD_REC_LID_SWITCH 0
	gpiocfg shows "GPIO1_GPIO10: read 0 drive 0"
        gpioset CCD_REC_LID_SWITCH 1
	gpiocfg doesn't show GPIO1_GPIO10 as an output

Change-Id: If74385135a572e7e5d0763fad9f5368fdec8d7a0
Signed-off-by: Mary Ruthven <mruthven@chromium.org>