dedede: Add z-state support
Dedede has support for the new low power "Z-state".  In this state. the
EC and H1 are unpowered, but power will be restored to the EC once one
of the wakeup events occurs.  These events are ACOK, lid open, and a
power button press.  This commit simply enables the Z-state when the EC
hibernates.

BUG=b:147819424
BRANCH=None
TEST=build and flash waddledoo, enter `hibernate`, verify that EC power
is turned off and can be restored by pressing the power button or
plugging in a charger.

Change-Id: I4f93efd0632f457354f4bf6bf0274b19a9cd799c
Signed-off-by: Aseda Aboagye <aaboagye@google.com>